add temp file while writing to device; improve stability

+++ KidPadTool/threadAction.h
@@ -6,21 +6,21 @@ CString m_driveNANDName;
 
 typedef struct _tagTRANSFER_PARAMETER_BLOCK
 {
-    TCHAR sourceFile[1024];
+	TCHAR sourceFile[1024];
 	TCHAR targetFile[1024];
 	CShockwaveflash1 *flash_pointer;
 } TRANSFER_PARAMETER_BLOCK;
 
 typedef struct _tagBIG_UI_DATA_PARAMETER_BLOCK
 {
-    TCHAR method[1024];
+	TCHAR method[1024];
 	TCHAR argument[1024];
 	CShockwaveflash1 *flash_pointer;
 } BIG_UI_DATA_PARAMETER_BLOCK;
 
 typedef struct _tagGENERAL_PARAMETER_BLOCK
 {
-    TCHAR parameter1[1024];
+	TCHAR parameter1[1024];
 	TCHAR parameter2[1024];
 	TCHAR parameter3[1024];
 	CShockwaveflash1 *flash_pointer;
@@ -35,7 +35,7 @@ void __cdecl threadGetDownloadedApps(void * p)
 	CFileFind finder;
 	TCHAR fileSizeBuf[16];
 	BOOL found = finder.FindFile(downloadDirectory + _T("*.npk"));
-	
+
 	while(found)
 	{
 		found = finder.FindNextFile();
@@ -111,69 +111,125 @@ void __cdecl threadUpdatePercentage(void * p)
 {
 	TRANSFER_PARAMETER_BLOCK *tpb = (TRANSFER_PARAMETER_BLOCK *)p;
 	CShockwaveflash1 *flash_pointer = tpb->flash_pointer;
-	
-	int err, err2;
-	CString target = CString(tpb->targetFile);
-	err = fsOpenFile((CHAR *)(LPCWSTR)target, NULL, O_CREATE|O_TRUNC );
-	
-	BYTE buffer[1024*8];
+
+	BYTE buffer[api_pointer->file_buff_size] = { 0 };
+	TCHAR buff01[20] = {0};
 	CFile file;
+	CString target = _T(""), tmpfile = _T("");
+	int hdl, status;
+	DWORD totalBytes =  0 ;
+	int completePercentage = 0, pre_completePercentage = 0;
+	DWORD dwBytesRemaining = 0;
+	INT nWriteCnt = 0;
+	UINT nBytesRead = 0;
+	CString str = _T("");
+	float fv = 0.00 ;
+
+	if(api_pointer->IsConnected(false) == false)
+	{
+		MessageBox(api_pointer->ownerWindow->m_hWnd, _T("设备处于断开状态， 文件传输失败。\r\n请先连接上设备， 然后再重试！"), _T("错误"), MB_OK|MB_ICONSTOP );
+		goto lb_exit ;
+	}
+
+
+	target = CString(tpb->targetFile);
+	tmpfile = target ;
+	tmpfile.Append(_T(".tmp"));
+	hdl = fsOpenFile((CHAR *)(LPCWSTR)tmpfile, NULL, O_CREATE|O_TRUNC );
+	if(hdl < 0)
+	{
+		::OutputDebugString(_T("打开设备上的文件["));
+		::OutputDebugString((LPCWSTR)target);
+		::OutputDebugString(_T("]失败！\r\n"));
+		goto lb_exit ; 
+	}
+	::OutputDebugString(_T("打开设备上的文件["));
+	::OutputDebugString((LPCWSTR)target);
+	::OutputDebugString(_T("]成功！\r\n"));
+
 	try
 	{
 		CFileException ex;
-		if ( !file.Open(CString(tpb->sourceFile), CFile::modeRead | CFile::shareDenyWrite, &ex) )
+		if (file.Open(CString(tpb->sourceFile), CFile::modeRead | CFile::shareDenyWrite, &ex) == FALSE )
 		{
 			TCHAR szError[1024];
 			ex.GetErrorMessage(szError, 1024);
-			return;
+			::OutputDebugString(szError);
+			::OutputDebugString(_T("\r\n"));
+			goto lb_exit ;
 		}
 
-		DWORD totalBytes = (DWORD)file.GetLength();
-		int completePercentage = 0;
-		DWORD dwBytesRemaining = (DWORD)file.GetLength();
+		totalBytes = (DWORD)file.GetLength();
+		completePercentage = 0;
+		dwBytesRemaining = totalBytes;
 
-		INT nWriteCnt;
-		UINT nBytesRead;
-		while ( dwBytesRemaining )
+		while ( dwBytesRemaining > 0 )
 		{
-			memset(buffer, 0, 1024*8);
-			nBytesRead = file.Read( buffer, sizeof(buffer) );
-			err2 = fsWriteFile( err, buffer, nBytesRead, &nWriteCnt );
-			if( err2 != FS_OK )
+			memset(buffer, 0, api_pointer->file_buff_size);
+			nBytesRead = file.Read(buffer, api_pointer->file_buff_size);
+			status = fsWriteFile(hdl, buffer, nBytesRead, &nWriteCnt );
+			if( status != FS_OK )
 			{
 				file.Close();
-				err2 = fsCloseFile( err );
-				return;
+				status = fsCloseFile( hdl );
+				hdl = 0 ;
+				goto lb_exit ;
 			}
 			dwBytesRemaining -= nBytesRead;
+			if(nWriteCnt != (INT)nBytesRead)
+			{
+				file.Close();
+				status = fsCloseFile( hdl );
+				hdl = 0 ;
+				goto lb_exit ;
+			}
 
 			// transfer percentage
-			completePercentage = ((totalBytes-dwBytesRemaining)*100)/totalBytes;
-			if (totalBytes > 20*1024*1024)
-				completePercentage = ((totalBytes-dwBytesRemaining)/10)/(totalBytes/1000);
-			CString str;
-			str.Format(L"%d", completePercentage);
-			flash_pointer->CallFunction(_T("<invoke name='FL_setTransferPercentage'><arguments><string>") + str + _T("</string></arguments></invoke>"));
-			//str.Format(L"totalBytes=%d, remainingBytes=%d, completed:%d", totalBytes, dwBytesRemaining, completePercentage);
-			//TRACE(str);
-			
-			CString out;
-			out.Format(L"total=%d, complete=%d, remain=%d, complete%%=%d%%\n", totalBytes, (totalBytes-dwBytesRemaining), dwBytesRemaining, completePercentage);
-			::OutputDebugString(out);
+			fv = (float)dwBytesRemaining  / (float)totalBytes ;
+			fv *= 100 ;
+			completePercentage = 100 - (int)fv;
+			if( completePercentage > pre_completePercentage)
+			{
+				memset(buff01, 0, 20);
+				::_itot(completePercentage, buff01, 10);
+				str = _T("<invoke name='FL_setTransferPercentage'><arguments><string>");
+				str.Append(buff01);
+				str.Append(_T("</string></arguments></invoke>"));
+				flash_pointer->CallFunction((LPCTSTR)str);
+				pre_completePercentage = completePercentage;
+#ifdef _DEBUG
+				if((completePercentage % 10) == 0)
+				{
+					CString out;
+					out.Format(L"total=%d, complete=%d, remain=%d, complete%%=%d%%\n", totalBytes, (totalBytes-dwBytesRemaining), dwBytesRemaining, completePercentage);
+					::OutputDebugString(out);
+				}
+#endif
+			}
 		}
 		file.Close();
 	}
 	catch ( CFileException* pEx )
 	{
-		err2 = fsCloseFile( err );
+		status = fsCloseFile( hdl );
+		hdl = 0 ;
 		pEx->Delete();
-		return;
+		fsDeleteFile((CHAR*)(LPCWSTR)tmpfile, NULL);
+		goto lb_exit ;
+	}
+	if(hdl > 0)	{ fsCloseFile( hdl ); }
+
+	if(api_pointer->WriteStream2DeviceFile(target, 0, 0, true) > 0)
+	{
+		TRACE(_T("COMPLETE TRANSFER\n"));
+		flash_pointer->CallFunction(_T("<invoke name='FL_completeTransfer'><arguments><string></string></arguments></invoke>"));
+
 	}
 
-	fsCloseFile( err );
 
-	TRACE(_T("COMPLETE TRANSFER\n"));
-	flash_pointer->CallFunction(_T("<invoke name='FL_completeTransfer'><arguments><string></string></arguments></invoke>"));
+lb_exit:
+
+	if(hdl > 0)	{ fsCloseFile( hdl ); }
 
 	free(p);
 	_endthread();
@@ -183,7 +239,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 {
 	WIN32_FIND_DATA info;
 	HANDLE hwnd;
-
+	CString ret ;
 	CString tempName;
 	CString tempName1;
 	CString srcTempName;
@@ -197,7 +253,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 		if (err != ERR_DIR_BUILD_EXIST)
 		{
 			MessageBox(api_pointer->ownerWindow->m_hWnd, _T("在设备上创建\"") + destName + _T("\"文件夹失败，可能已经存在或设备磁盘空间已满"), _T("错误"), MB_OK|MB_ICONSTOP );
-			flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+			ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 			return FALSE;
 		}
 	}
@@ -231,7 +287,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 				{
 					FindClose( hwnd );
 					MessageBox(api_pointer->ownerWindow->m_hWnd, _T("磁盘已满"), _T("错误"), MB_OK|MB_ICONSTOP);
-					flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+					ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 					return FALSE;
 				}
 			}
@@ -243,11 +299,11 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 			{
 				FindClose( hwnd );
 				MessageBox(api_pointer->ownerWindow->m_hWnd, _T("在设备上创建\"") + suFileName + _T("\"文件失败，可能已经存在或设备磁盘空间已满"), _T("错误"), MB_OK|MB_ICONSTOP);
-				flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+				ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 				return FALSE;
 			}
 
-			BYTE buffer[4096];
+			BYTE buffer[api_pointer->file_buff_size] = {0};
 			CFile file;
 			try
 			{
@@ -256,7 +312,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 					err2 = fsCloseFile( err );
 					FindClose( hwnd );
 					MessageBox(api_pointer->ownerWindow->m_hWnd, srcTempName, _T("打开计算机上文件失败"), MB_OK|MB_ICONSTOP );
-					flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+					ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 					return FALSE;
 				}
 
@@ -269,7 +325,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 				UINT nBytesRead;
 				while ( dwBytesRemaining )
 				{
-					nBytesRead = file.Read( buffer, sizeof(buffer) );
+					nBytesRead = file.Read( buffer, api_pointer->file_buff_size );
 					err2 = fsWriteFile( err, buffer, nBytesRead, &nWriteCnt );
 					if( err2 != FS_OK )
 					{
@@ -277,7 +333,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 						FindClose( hwnd );
 						err2 = fsCloseFile( err );
 						MessageBox(api_pointer->ownerWindow->m_hWnd, _T("在设备上写\"") + suFileName + _T("\"文件失败，可能设备磁盘空间已满"), _T("错误"), MB_OK|MB_ICONSTOP );
-						flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+						CString ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 						return FALSE;
 					}
 					dwBytesRemaining -= nBytesRead;
@@ -291,7 +347,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 					arg += info.cFileName;
 					arg += str;
 					arg += _T("</string></arguments></invoke>");
-					flash_pointer->CallFunction(arg);
+					ret = flash_pointer->CallFunction(arg);
 				}
 				file.Close();
 			}
@@ -301,7 +357,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 				FindClose( hwnd );
 				MessageBox(api_pointer->ownerWindow->m_hWnd, pEx->m_strFileName, _T("Get length, read, or close PC file failed"), MB_OK|MB_ICONSTOP );
 				pEx->Delete();
-				flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+				ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 				return FALSE;
 			}
 
@@ -310,7 +366,7 @@ BOOL CopyDirectory(CString srcName, CString destName, CShockwaveflash1 *flash_po
 			{
 				FindClose( hwnd );
 				MessageBox(api_pointer->ownerWindow->m_hWnd, _T("Close \"") + suFileName + _T("\" failed, may disk full"), _T("Error"), MB_OK|MB_ICONSTOP );
-				flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
+				ret = flash_pointer->CallFunction(_T("<invoke name='") + failCommand + _T("'><arguments><string></string></arguments></invoke>"));
 				return FALSE;
 			}
 		}
@@ -328,9 +384,16 @@ void __cdecl threadCopyPcDirectoryToDevice(void * p)
 	CString srcName(tpb->parameter1);
 	CString destName(tpb->parameter2);
 
-	if (CopyDirectory(srcName, destName, flash_pointer, _T("FL_installSetTransferInformation"), _T("FL_failInstall")))
+	if(api_pointer->IsConnected(false) == false)
 	{
-		flash_pointer->CallFunction(_T("<invoke name='FL_installCompleteTransfer'><arguments><string></string></arguments></invoke>"));
+		MessageBox(api_pointer->ownerWindow->m_hWnd, _T("设备处于断开状态， 文件传输失败。\r\n请先连接上设备， 然后再重试！"), _T("错误"), MB_OK|MB_ICONSTOP );
+	}
+	else
+	{
+		if (CopyDirectory(srcName, destName, flash_pointer, _T("FL_installSetTransferInformation"), _T("FL_failInstall")) == TRUE)
+		{
+			CString ret = flash_pointer->CallFunction(_T("<invoke name='FL_installCompleteTransfer'><arguments><string></string></arguments></invoke>"));
+		}
 	}
 
 	free(p);
@@ -343,12 +406,18 @@ void __cdecl threadInstallBuiltIn(void * p)
 	CShockwaveflash1 *flash_pointer = tpb->flash_pointer;
 	CString srcParent(tpb->parameter1);
 	CString destParent(tpb->parameter2);
-	
+
+	if(api_pointer->IsConnected(false) == false)
+	{
+		MessageBox(api_pointer->ownerWindow->m_hWnd, _T("设备处于断开状态， 文件传输失败。\r\n请先连接上设备， 然后再重试！"), _T("错误"), MB_OK|MB_ICONSTOP );
+		goto lb_exit ;
+	}
+
 	if (CopyDirectory(srcParent+_T("builtIn"), destParent+_T("builtIn2"), flash_pointer, _T("FL_updateSetTransferInformation"), _T("FL_failUpdate")))
 	{
 		api_pointer->DeletePcDirectory(srcParent+_T("builtIn"));
 		::RemoveDirectory(srcParent+_T("builtIn"));
-		
+
 		// delete obsolete builtIn on device
 		api_pointer->DeleteDirectoryOnDevice(destParent+_T("builtIn"));
 
@@ -362,7 +431,7 @@ void __cdecl threadInstallBuiltIn(void * p)
 	{
 		api_pointer->DeletePcDirectory(srcParent+_T("base"));
 		::RemoveDirectory(srcParent+_T("base"));
-		
+
 		// delete obsolete base on device
 		api_pointer->DeleteDirectoryOnDevice(destParent+_T("base"));
 
@@ -371,6 +440,8 @@ void __cdecl threadInstallBuiltIn(void * p)
 		if (fsRenameFile((CHAR *)(LPCWSTR)(destParent+_T("base2")), NULL, (CHAR *)(LPCWSTR)(destParent+_T("base")), NULL, TRUE) == FS_OK)
 			flash_pointer->CallFunction(_T("<invoke name='FL_updateCompleteTransfer'><arguments><string></string></arguments></invoke>"));
 	}
+
+lb_exit:
 
 	free(p);
 	_endthread();
